fix: rename cmd and improve tests
